cardiac-care-ragcot_conf/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Harry5174", email = "harisjaved010@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi[standard]>=0.115.11",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-google-genai>=2.1.2",
    "langchain-openai>=0.3.9",
    "openai>=1.67.0",
    "packaging>=24.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "rouge-score>=0.1.2",
    "sentence-transformers>=3.4.1",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
